microservicios/catalog_service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
PyJWT==2.8.0
dicttoxml==1.7.4
requests==2.31.0
